requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
psutil==5.9.4
python-telegram-bot==20.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
speedtest-cli==2.1.3
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
